nightlies: trim version from artifact name in latest tag [deploy]
The `latest-` family of tags was introduced to simplify automation by providing
a mean to fetch the latest nightly built without meddling with Github API.
However this doesn't work out perfectly in practice due to the version
of the tarball is still a part of the link.

This commit strips the version out for artifacts in latest- tags to solve that.